Refresh OpOverloadPacket if a new OpOverload gets added
If a user accesses an OpOverloadPacket, then creates a new OpOverload,
then uses the OpOverloadPacket, the new OpOverload never gets hit. This
is because OpOverloadPacket caches OpOverloads when it is constructed.

This PR fixes the problem by "refreshing" the OpOverloadPacket if a new
OpOverload gets constructed and the OpOverloadPacket exists.

Test Plan:
- new tests

ghstack-source-id: 824de3c5d154eef75626f5d0227d9fb478c0ecfe
Pull Request resolved: https://github.com/pytorch/pytorch/pull/124654